backend/requirements.txt
django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
twilio==9.10.2
djangorestframework-simplejwt==5.5.1
